Refactoring the prim standard functions (#122)
* unify type construction patterns in the stdlib

* split out finalizers to a separate module

* split out IR building facilites from stdlib

* split out domain-modules from stdlib
